requirements.txt
kafka-python==2.0.2
numpy==2.2.2
opencv-python==4.11.0.86
python-dotenv==1.0.1
requests==2.32.3